feat(core-lib): 新增 Exception 构建程序。
